Another rationale for loaders.
Keep one thread by default until the queue is filled and expand to 4 so that it does not overload the connection pool for database (SQLiteConnectionPool) which is 4 on firestick4k.
Do not have a too large LinkedBlockingQueue (default is 10 and 5200 was too much).
Having 4 threads speeds things up in general (when queue is full) and 256 is enough headroom in queue capacity to avoid epileptic effect on VideoByFragments (IMHO) or at least a good compromise.
Explanation: since it starts with a corePoolSize of 1, it is mono-threaded until queue is filled and there "usually" it is not seen...
Reducing keepAlive to 1s instead of 10s is perhaps questionable (TBC).

All this has been tested visually to try to optimize things.
MoviesByGenre is really slow but tweaking these parameters does not help.